Move into GridField folder, add test
